Fix: - Assegnato lo stato di invio al messaggio, lo stato viene aggiornato ad ogni invio di un pacchetto. - onMessageSent è stato legato al singolo messaggio

+++ sms/src/test/java/com/gruppo4/sms/SMSControllerUnitTest.java
@@ -10,7 +10,7 @@ public class SMSControllerUnitTest {
 
     @Before
     public void init() {
-        smsController = new SMSController((byte)1);
+        smsController = new SMSController(1);
     }
 
     @Test
